Fix org to html exporting githook

+++ bashscripts/mirror_org_as_html.sh
@@ -4,29 +4,38 @@ set -x
 
 echo "$(env | grep PATH)"
 
+# Loop through each staged .org file
 git diff --cached --name-only --diff-filter=AM | grep '.org$' | while IFS= read -r org_file; do
-    echo "Processing: $org_file"
-    # base_name=$(basename "${org_file}")
     org_file_dir=$(dirname "${org_file}")
-    # echo "$(pwd)"
-    # cd "$git_root" || exit
     git_root=$(git rev-parse --show-toplevel)
     relative_path=$(realpath --relative-to="$git_root" "$org_file_dir")
     html_folder="${git_root}/html/${relative_path}"
-
     [[ -d "$html_folder" ]] || mkdir -p "$html_folder"
 
-    # Use Emacs in batch mode to export the Org file to HTML
     emacs --batch \
         --eval "(progn
                 (require 'org)
                 (find-file \"$org_file\")
                 (org-html-export-to-html))"
-    # export-org-to-html "${org_file}"
 
     html_file="${org_file%.org}.html"
+
+    if [[ ! -f "$html_file" ]]; then
+        echo "Error: HTML file not generated: $html_file"
+        exit
+    fi
+
     mv "$html_file" "$html_folder"
-    cp "${dirname}/*" "${html_folder}"
-    git add "html/${html_file}"
+
+    # Create symbolic links for everything in the Org file's directory in the HTML folder
+    # this is so that images and so on render
+    find "$org_file_dir" -maxdepth 1 -mindepth 1 | while IFS= read -r file; do
+        base_file=$(basename "$file")
+        relative_path=$(realpath --relative-to="$html_folder" "$git_root/$file")
+        # shenanigans with relative paths to get symlinks to work
+        ln -Ts "$relative_path" "$html_folder/$base_file"
+    done
+
+    git add --all "$html_folder"
 
 done
